Modifited FinanceService, added methods for authorization(IsAccountOwnerOfFinanceOperationAsync, IsAccountOwnerOfWalletAsync)

+++ Finance manager/ApplicationLayer/Controllers/FinanceOperationTypeController.cs
@@ -11,13 +11,13 @@ public class FinanceOperationTypeController : BaseController
 {
     private readonly IFinanceService _financeService;
 
-    public FinanceOperationTypeController(IFinanceService financeService, IMapper mapper, ILogger<BaseController> logger) : base(mapper, logger)
+    public FinanceOperationTypeController(IFinanceService financeService, IMapper mapper, ILogger<FinanceOperationTypeController> logger) : base(mapper, logger)
     {
         _financeService = financeService ?? throw new ArgumentNullException(nameof(financeService));
     }
 
     [HttpGet("types/wallet/{walletId}")]
-    public async Task<List<FinanceOperationTypeDTO>> GetAllAsync(int walletId)
+    public async Task<IActionResult> GetAllAsync(int walletId)
     {
         _logger.LogInformation("GetAllAsync called to retrieve finance operation types for wallet Id: {WalletId}", walletId);
 
@@ -27,11 +27,11 @@ public class FinanceOperationTypeController : BaseController
 
         _logger.LogInformation("{Count} finance operation types retrieved for wallet Id: {WalletId}", operationTypes.Count, walletId);
 
-        return operationTypes;
+        return Ok(operationTypes);
     }
 
     [HttpPost("create")]
-    public async Task<FinanceOperationTypeDTO> AddAsync([FromBody] FinanceOperationTypeDTO dto)
+    public async Task<IActionResult> AddAsync([FromBody] FinanceOperationTypeDTO dto)
     {
         _logger.LogInformation("AddAsync called to add a new finance operation type with name: {Name}", dto.Name);
 
@@ -41,11 +41,11 @@ public class FinanceOperationTypeController : BaseController
 
         _logger.LogInformation("Finance operation type with Id: {Id} added successfully", newOperationType.Id);
 
-        return newOperationType;
+        return Ok(newOperationType);
     }
 
     [HttpPut("update")]
-    public async Task<FinanceOperationTypeDTO> UpdateAsync([FromBody] FinanceOperationTypeDTO dto)
+    public async Task<IActionResult> UpdateAsync([FromBody] FinanceOperationTypeDTO dto)
     {
         _logger.LogInformation("UpdateAsync called to update finance operation type with Id: {Id}", dto.Id);
 
@@ -55,17 +55,22 @@ public class FinanceOperationTypeController : BaseController
 
         _logger.LogInformation("Finance operation type with Id: {Id} updated successfully", updatedOperationType.Id);
 
-        return updatedOperationType;
+        return Ok(updatedOperationType);
     }
 
     [HttpDelete("remove/{id}")]
-    public async Task DeleteAsync(int id)
+    public async Task<IActionResult> DeleteAsync(int id)
     {
         _logger.LogInformation("DeleteAsync called to remove finance operation type with Id: {Id}", id);
+
+        if (!(await _financeService.IsAccountOwnerOfFinanceOperationTypeAsync(GetUserId(), id)))
+            throw new UnauthorizedAccessException();
 
         await _financeService.DeleteFinanceOperationTypeAsync(id);
 
         _logger.LogInformation("Finance operation type with Id: {Id} deleted successfully", id);
+
+        return Ok();
     }
 
 }
